Refactoring addressbook synchronization
