+++ src/Presentation/Async/Message/SyncAddressBook.php
@@ -1,0 +1,12 @@
+<?php
+
+/**
+ * @author: Julien Mercier-Rojas <julien@jeckel-lab.fr>
+ * Created at: 07/02/2024
+ */
+
+declare(strict_types=1);
+
+namespace App\Presentation\Async\Message;
+
+final readonly class SyncAddressBook {}
